requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==45.0.6
cssselect2==0.8.0
Django==4.2.7
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
html5lib==1.1
idna==3.10
kombu==5.5.4
lxml==6.0.0
oscrypto==1.3.0
packaging==25.0
phonenumbers==8.13.23
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.9.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
